+++ example-client/views/rest-test/_form.php
@@ -1,0 +1,25 @@
+<?php
+
+use yii\helpers\Html;
+use yii\bootstrap\ActiveForm;
+
+/* @var $this yii\web\View */
+/* @var $model app\models\Test */
+/* @var $form yii\bootstrap\ActiveForm */
+?>
+
+<div class="account-form">
+
+    <?php $form = ActiveForm::begin(); ?>
+
+    <?= $form->field($model, 'name')->textInput() ?>
+    <?= $form->field($model, 'status')->textInput() ?>
+
+    <div class="form-group">
+        <?= Html::submitButton($model->isNewRecord ? 'Добавить' : 'Обновить',
+            ['class' => $model->isNewRecord ? 'btn btn-success' : 'btn btn-primary']) ?>
+    </div>
+
+    <?php ActiveForm::end(); ?>
+
+</div>