+++ example-client/views/rest-test/index.php
@@ -1,0 +1,34 @@
+<?php
+
+use yii\helpers\Url;
+use yii\helpers\Html;
+use yii\grid\GridView;
+
+/* @var $this yii\web\View */
+/* @var $searchModel app\models\TestSearch */
+/* @var $dataProvider yii\restclient\RestDataProvider */
+
+$this->title = 'Список';
+$this->params['breadcrumbs'][] = $this->title;
+?>
+<div class="account-index">
+
+    <h1><?= Html::encode($this->title) ?></h1>
+
+    <br/>
+    <p>
+        <?= Html::a('Добавить', ['create'], ['class' => 'btn btn-success']) ?>
+    </p>
+
+    <?= GridView::widget([
+        'dataProvider' => $dataProvider,
+        'filterModel' => $searchModel,
+        'columns' => [
+            ['class' => 'yii\grid\SerialColumn'],
+            'name',
+            'status',
+            ['class' => 'yii\grid\ActionColumn'],
+        ],
+    ]); ?>
+
+</div>