+++ example-client/views/rest-test/view.php
@@ -1,0 +1,37 @@
+<?php
+
+use yii\helpers\Html;
+use yii\widgets\DetailView;
+
+/* @var $this yii\web\View */
+/* @var $model app\models\Test */
+
+$this->title = $model->id;
+$this->params['breadcrumbs'][] = ['label' => 'Счета', 'url' => ['index']];
+$this->params['breadcrumbs'][] = $this->title;
+?>
+<div class="account-view">
+
+    <h1><?= Html::encode($this->title) ?></h1>
+
+    <p>
+        <?= Html::a('Обновить', ['update', 'id' => $model->id], ['class' => 'btn btn-primary']) ?>
+        <?= Html::a('Удалить', ['delete', 'id' => $model->id], [
+            'class' => 'btn btn-danger',
+            'data' => [
+                'confirm' => 'Вы уверены что хотите удалить запись?',
+                'method' => 'post',
+            ],
+        ]) ?>
+    </p>
+
+    <?= DetailView::widget([
+        'model' => $model,
+        'attributes' => [
+            'id',
+            'name',
+            'status',
+        ],
+    ]) ?>
+
+</div>